requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
Authlib==1.4.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
entrypoints==0.4
extra-streamlit-components==0.1.71
Faker==37.0.0
favicon==0.7.0
fonttools==4.56.0
fpdf==1.7.2
gitdb==4.0.12
GitPython==3.1.44
groq==0.18.0
h11==0.14.0
htbuilder==0.9.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain-core==0.3.34
langchain-groq==0.2.4
langsmith==0.3.6
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
narwhals==1.25.2
numpy==2.2.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
prometheus_client==0.21.1
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pymdown-extensions==10.14.3
pymongo==4.11.2
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.3.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
st-annotated-text==4.0.2
st-chat-message==0.3.8
st-theme==1.2.3
streamlit==1.42.0
streamlit-autorefresh==1.0.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-cookies-controller==0.0.4
streamlit-embedcode==0.1.2
streamlit-extras==0.5.5
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-navigation-bar==3.3.0
streamlit-option-menu==0.4.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
zstandard==0.23.0
